Optimal Layouts

Optimal Layouts

Overview

In DataKnots.jl, we often need to visualize composite data structures or complex Julia expressions. For this purpose, module DataKnots.Layouts implements a pretty-printing engine.

using DataKnots.Layouts

To format a data structure, we need to encode its possible layouts in the form of a layout expression.

A fixed single-line layout is created with Layouts.literal().

Layouts.literal("department")
#-> literal("department")

Layouts could be combined using horizontal and vertical composition operators.

lhz = Layouts.literal("department") * Layouts.literal(".") * Layouts.literal("name")
#-> literal("department.name")

lvt = Layouts.literal("department") / Layouts.literal("name")
#-> literal("department") / literal("name")

Function Layouts.pretty_print() serializes the layout.

pretty_print(lhz)
#-> department.name

pretty_print(lvt)
#=>
department
name
=#

To indicate that we can choose between several different layouts, use the choice operator.

l = lhz | lvt
#-> literal("department.name") | literal("department") / literal("name")

The pretty-printing engine can search through possible layouts to find the best fit, which is expressed as a layout expression without a choice operator.

Layouts.best(Layouts.fit(l))
#-> literal("department.name")

The module implements the optimal layout algorithm described in https://research.google.com/pubs/pub44667.html.

API Reference

DataKnots.Layouts.pretty_print — Function.

Layouts.pretty_print([io::IO], data)

Formats the data so that it fits the width of the output screen.

source

DataKnots.Layouts.print_code — Function.

Layouts.print_code([io::IO], code)

Formats a Julia expression.

source

Test Suite

We start with creating a simple tree structure.

struct Node
    name::Symbol
    arms::Vector{Node}
end

Node(name) = Node(name, [])

tree =
    Node(:a, [Node(:an, [Node(:anchor, [Node(:anchorage), Node(:anchorite)]),
                           Node(:anchovy),
                           Node(:antic, [Node(:anticipation)])]),
               Node(:arc, [Node(:arch, [Node(:archduke), Node(:archer)])]),
               Node(:awl)])
#-> Node(:a, Main.layouts.md.Node[ … ])

To specify a layout expression for Node objects, we need to override Layout.tile(). Layout expressions are assembled from Layouts.literal() primitives using operators * (horizontal composition), / (vertical composition), and | (choice).

function Layouts.tile(tree::Node)
    if isempty(tree.arms)
        return Layouts.literal("Node($(repr(tree.name)))")
    end
    arm_lts = [Layouts.tile(arm) for arm in tree.arms]
    v_lt = h_lt = nothing
    for (k, arm_lt) in enumerate(arm_lts)
        if k == 1
            v_lt = arm_lt
            h_lt = Layouts.nobreaks(arm_lt)
        else
            v_lt = v_lt * Layouts.literal(",") / arm_lt
            h_lt = h_lt * Layouts.literal(", ") * Layouts.nobreaks(arm_lt)
        end
    end
    return Layouts.literal("Node($(repr(tree.name)), [") *
           (v_lt | h_lt) *
           Layouts.literal("])")
end

Now we can use function pretty_print() to render a nicely formatted representation of the tree.

pretty_print(stdout, tree)
#=>
Node(:a, [Node(:an, [Node(:anchor, [Node(:anchorage), Node(:anchorite)]),
                     Node(:anchovy),
                     Node(:antic, [Node(:anticipation)])]),
          Node(:arc, [Node(:arch, [Node(:archduke), Node(:archer)])]),
          Node(:awl)])
=#

We can control the width of the output.

pretty_print(IOContext(stdout, :displaysize => (24, 60)), tree)
#=>
Node(:a, [Node(:an, [Node(:anchor, [Node(:anchorage),
                                    Node(:anchorite)]),
                     Node(:anchovy),
                     Node(:antic, [Node(:anticipation)])]),
          Node(:arc, [Node(:arch, [Node(:archduke),
                                   Node(:archer)])]),
          Node(:awl)])
=#

We can display the layout expression itself, both the original and the optimized variants.

Layouts.tile(tree)
#=>
literal("Node(:a, [")
* (literal("Node(:an, [")
   * (literal("Node(:anchor, [")
   ⋮
=#

Layouts.best(Layouts.fit(stdout, Layouts.tile(tree)))
#=>
literal("Node(:a, [")
* (literal("Node(:an, [")
   * (literal("Node(:anchor, [Node(:anchorage), Node(:anchorite)]),")
   ⋮
=#

For some built-in data structures, automatic layout is already provided.

data = [
    (name = "RICHARD A", position = "FIREFIGHTER", salary = 90018),
    (name = "DEBORAH A", position = "POLICE OFFICER", salary = 86520),
    (name = "KATHERINE A", position = "PERSONAL COMPUTER OPERATOR II", salary = 60780)
]

pretty_print(data)
#=>
[(name = "RICHARD A", position = "FIREFIGHTER", salary = 90018),
 (name = "DEBORAH A", position = "POLICE OFFICER", salary = 86520),
 (name = "KATHERINE A",
  position = "PERSONAL COMPUTER OPERATOR II",
  salary = 60780)]
=#

This includes Julia syntax trees.

Q = :(
    Employee
    >> ThenFilter(Department >> Name .== "POLICE")
    >> ThenSort(Salary >> Desc)
    >> ThenSelect(Name, Position, Salary)
    >> ThenTake(10)
)

print_code(Q)
#=>
Employee
>> ThenFilter(Department >> Name .== "POLICE")
>> ThenSort(Salary >> Desc)
>> ThenSelect(Name, Position, Salary)
>> ThenTake(10)
=#