Reference

Reference

DataKnots.DataKnot — Type.

DataKnot(cell::AbstractVector, shp::AbstractShape)

Encapsulates a data cell serialized in a column-oriented form.

source

DataKnots.Count — Method.

Count(X) :: Query
Each(X >> Count) :: Query

Counts the number of elements produced by the query.

source

DataKnots.Drop — Method.

Drop(N) :: Query

Drops the first N elements of the input, keeps the rest.

source

DataKnots.Each — Method.

Each(X) :: Query

This query evaluates X elementwise.

source

DataKnots.Filter — Method.

Filter(X) :: Query

Filters the input by the given condition.

source

DataKnots.Get — Method.

Get(name) :: Query

This query emits the value of a record field.

source

DataKnots.Given — Method.

Given(X₁, X₂ … Xₙ, Q) :: Query

Evaluates the query with the given context parameters.

source

DataKnots.Keep — Method.

Keep(X₁, X₂ … Xₙ) :: Query

Defines context parameters.

source

DataKnots.Label — Method.

Label(lbl::Symbol) :: Query

Assigns a label to the output.

source

DataKnots.Lift — Method.

Lift(f, (X₁, X₂ … Xₙ)) :: Query

This query uses the outputs of X₁, X₂ … Xₙ as arguments of f. The output of f is emitted.

source

DataKnots.Lift — Method.

Lift(val) :: Query

This query emits the given value.

source

DataKnots.Max — Method.

Max(X) :: Query
Each(X >> Max) :: Query

Finds the maximum among the elements produced by the query.

source

DataKnots.Min — Method.

Min(X) :: Query
Each(X >> Min) :: Query

Finds the minimum among the elements produced by the query.

source

DataKnots.Record — Method.

Record(X₁, X₂ … Xₙ) :: Query

This query emits records, whose fields are generated by X₁, X₂ … Xₙ.

source

DataKnots.Sum — Method.

Sum(X) :: Query
Each(X >> Sum) :: Query

Sums the elements produced by the query.

source

DataKnots.Tag — Method.

Tag(name::Symbol, F) :: Query
Tag(name::Symbol, (X₁, X₂ … Xₙ), F) :: Query

Assigns a name to a query.

source

DataKnots.Take — Method.

Take(N) :: Query

Keeps the first N elements of the input, drops the rest.

source