Query Algebra

Query Algebra

Overview

In DataKnots, structured data is stored in a column-oriented format, serialized using specialized composite vector types. Consequently, operations on data must also be adapted to the column-oriented format.

Module DataKnots implements the Query interface of vectorized data transformations and provives a rich library of query primitives and combinators.

using DataKnots:
    @VectorTree,
    as_block,
    chain_of,
    column,
    decode_missing,
    decode_tuple,
    decode_vector,
    flat_block,
    in_block,
    in_tuple,
    lift,
    lift_block,
    lift_const,
    lift_null,
    lift_to_block,
    lift_to_block_tuple,
    lift_to_tuple,
    pass,
    pull_block,
    pull_every_block,
    tuple_of

Lifting

Lifting lets us convert a scalar function to a query.

Any unary scalar function could be lifted to a vectorized form. Consider, for example, function titlecase(), which transforms the input string by capitalizing the first letter of each word and converting every other character to lowercase.

titlecase("JEFFERY A")      #-> "Jeffery A"

This function can be converted to a query using the lift operator.

q = lift(titlecase)
q(["JEFFERY A", "JAMES A", "TERRY A"])
#-> ["Jeffery A", "James A", "Terry A"]

If a scalar function takes several arguments, it could be lifted to a query on TupleVector instances. For example, the comparison operator >, which maps a pair of integer values to a Boolean value, could be lifted to a query lift_to_tuple(>) that transforms a TupleVector instance with two integer columns to a Boolean vector.

q = lift_to_tuple(>)
q(@VectorTree (Int, Int) [260004 200000; 185364 200000; 170112 200000])
#-> Bool[true, false, false]

In a similar manner, a function with a vector argument can be converted to a query on BlockVector instances. For example, function length(), which returns the length of a vector, could be lifted to a query lift_to_block(length) that transforms a block vector to an integer vector containing block lengths.

q = lift_to_block(length)
q(@VectorTree [String] [["JEFFERY A", "NANCY A"], ["JAMES A"]])
#-> [2, 1]

A constant value could be lifted to a query as well. The lifted constant maps any input vector to a vector of constant values.

q = lift_const(200000)
q(["JEFFERY A", "JAMES A", "TERRY A"])
#-> [200000, 200000, 200000]

Query interface

Functions such as lift(), lift_to_tuple(), and many others return a Query object. The Query interface represents a vectorized data transformation that maps an input vector to an output vector of the same length.

Functions that take one or more Query instances as arguments and return a new Query object as the result are called combinators. Combinators are used to assemble elementary queries into complex query expressions.

For example, composition combinator chain_of() assembles a series of queries into a sequential composition, which transforms the input vector by sequentially applying the given queries.

q = chain_of(lift(split), lift(first), lift(titlecase))
q(["JEFFERY A", "JAMES A", "TERRY A"])
#-> ["Jeffery", "James", "Terry"]

Another combinator, tuple constructor tuple_of() assembles a series of queries into a parallel composition. It outputs a TupleVector instance, which columns are generated by applying the given queries to the input vector.

q = tuple_of(lift(titlecase), lift(last))
q(["JEFFERY A", "JAMES A", "TERRY A"])
#-> @VectorTree (String, Char) [("Jeffery A", 'A'), ("James A", 'A'), ("Terry A", 'A')]

An individual column of a TupleVector instance could be extracted using a column() query.

q = column(:salary)
q(@VectorTree (name=String, salary=Int) [("JEFFERY A", 101442), ("JAMES A", 103350), ("TERRY A", 93354)])
#-> [101442, 103350, 93354]

API Reference

DataKnots.Query — Type.

Query(op, args...)

A query object represents a vectorized data transformation.

Parameter op is a function that performs the transformation; args are extra arguments to be passed to the function.

The query transforms any input vector by invoking op with the following arguments:

op(rt::Runtime, input::AbstractVector, args...)

The result of op must be the output vector, which should be of the same length as the input vector.

source

DataKnots.Runtime — Type.

Runtime()

Runtime state for query evaluation.

source

DataKnots.any_block — Method.

any_block() :: Query

This query applies any to a block vector with Bool elements.

source

DataKnots.as_block — Method.

as_block() :: Query

This query produces a block vector with one-element blocks wrapping the values of the input vector.

source

DataKnots.chain_of — Function.

chain_of(q₁::Query, q₂::Query … qₙ::Query) :: Query

This query sequentially applies q₁, q₂ … qₙ.

source

DataKnots.column — Method.

column(lbl::Union{Int,Symbol}) :: Query

This query extracts the specified column of a tuple vector.

source

DataKnots.count_block — Method.

count_block() :: Query

This query converts a block vector to a vector of block lengths.

source

DataKnots.decode_missing — Method.

decode_missing() :: Query

This query transforms a vector that contains missing elements to a block vector with missing elements replaced by empty blocks.

source

DataKnots.decode_tuple — Method.

decode_tuple() :: Query

This query transforms a vector of tuples to a tuple vector.

source

DataKnots.decode_vector — Method.

decode_vector() :: Query

This query transforms a vector with vector elements to a block vector.

source

DataKnots.designate — Function.

designate(::Query, ::Signature) :: Query
designate(::Query, ::InputShape, ::OutputShape) :: Query
q::Query |> designate(::Signature) :: Query
q::Query |> designate(::InputShape, ::OutputShape) :: Query

Sets the query signature.

source

DataKnots.flat_block — Method.

flat_block() :: Query

This query flattens a nested block vector.

source

DataKnots.flat_tuple — Method.

flat_tuple(lbl::Union{Int,Symbol}) :: Query

This query flattens a nested tuple vector.

source

DataKnots.in_block — Method.

in_block(q::Query) :: Query

This query transforms a block vector by applying q to its vector of elements.

source

DataKnots.in_tuple — Method.

in_tuple(lbl::Union{Int,Symbol}, q::Query) :: Query

This query transforms a tuple vector by applying q to the specified column.

source

DataKnots.lift — Method.

lift(f) :: Query

f is any scalar unary function.

The query applies f to each element of the input vector.

source

DataKnots.lift_block — Function.

lift_block(block::AbstractVector, card::Cardinality) :: Query

This query produces a block vector filled with the given block.

source

DataKnots.lift_const — Method.

lift_const(val) :: Query

This query produces a vector filled with the given value.

source

DataKnots.lift_null — Method.

lift_null() :: Query

This query produces a block vector with empty blocks.

source

DataKnots.lift_to_block — Function.

lift_to_block(f) :: Query
lift_to_block(f, default) :: Query

f is a function that expects a vector argument.

The query applies f to each block of the input block vector. When a block is empty, default (if specified) is used as the output value.

source

DataKnots.lift_to_block_tuple — Method.

lift_to_block_tuple(f) :: Query

f is an n-ary function.

This query expects the input to be an n-tuple vector with each column being a block vector. The query produces a block vector, where each block is generated by applying f to every combination of values from the input blocks.

source

DataKnots.lift_to_tuple — Method.

lift_to_tuple(f) :: Query

f is an n-ary function.

The query applies f to each row of an n-tuple vector.

source

DataKnots.optimize — Method.

optimize(::Query) :: Query

Rewrites the query to make it (hopefully) faster.

source

DataKnots.pass — Method.

pass() :: Query

This query returns its input unchanged.

source

DataKnots.pull_block — Method.

pull_block(lbl::Union{Int,Symbol}) :: Query

This query transforms a tuple vector with a column of blocks to a block vector with tuple elements.

source

DataKnots.pull_every_block — Method.

pull_every_block() :: Query

This query transforms a tuple vector with block columns to a block vector with tuple elements.

source

DataKnots.sieve — Method.

sieve() :: Query

This query filters a vector of pairs by the second column. The query expects a pair vector, whose second column is a Bool vector. It produces a block vector with 0-element or 1-element blocks containing the elements of the first column.

source

DataKnots.signature — Method.

signature(::Query) :: Signature

Returns the query signature.

source

DataKnots.take_by — Function.

take_by(rev::Bool=false) :: Query

This query takes a pair vector of blocks and integers, and returns the first column with blocks restricted by the second column.

source

DataKnots.take_by — Function.

take_by(N::Int, rev::Bool=false) :: Query

This query transforms a block vector by keeping the first N elements of each block. If rev is true, the query drops the first N elements of each block.

source

DataKnots.tuple_of — Method.

tuple_of(q₁::Query, q₂::Query … qₙ::Query) :: Query

This query produces an n-tuple vector, whose columns are generated by applying q₁, q₂ … qₙ to the input vector.

source

Test Suite

Lifting

Many vector operations can be generated by lifting. For example, lift_const() generates a primitive operation that maps any input vector to the output vector of the same length filled with the given value.

q = lift_const(200000)
#-> lift_const(200000)

q(["GARRY M", "ANTHONY R", "DANA A"])
#-> [200000, 200000, 200000]

Similarly, the output of lift_block() is a block vector filled with the given block.

q = lift_block(["POLICE", "FIRE"])
#-> lift_block(["POLICE", "FIRE"], OPT | PLU)

q(["GARRY M", "ANTHONY R", "DANA A"])
#-> @VectorTree [String] [["POLICE", "FIRE"], ["POLICE", "FIRE"], ["POLICE", "FIRE"]]

A variant of lift_block() called lift_null() outputs a block vector with empty blocks.

q = lift_null()
q(["GARRY M", "ANTHONY R", "DANA A"])
#-> @VectorTree [Union{}, OPT] [missing, missing, missing]

Any scalar function could be lifted to a vector operation by applying it to each element of the input vector.

q = lift(titlecase)
#-> lift(titlecase)

q(["GARRY M", "ANTHONY R", "DANA A"])
#-> ["Garry M", "Anthony R", "Dana A"]

Similarly, any scalar function of several arguments could be lifted to an operation on tuple vectors.

q = lift_to_tuple(>)
#-> lift_to_tuple(>)

q(@VectorTree (Int, Int) [260004 200000; 185364 200000; 170112 200000])
#-> Bool[true, false, false]

It is also possible to apply a scalar function of several arguments to a tuple vector that has block vectors for its columns. In this case, the function is applied to every combination of values from all the blocks on the same row.

q = lift_to_block_tuple(>)

q(@VectorTree ([Int], [Int]) [[260004, 185364, 170112] 200000; missing 200000; [202728, 197736] [200000, 200000]])
#-> @VectorTree [Bool] [[true, false, false], [], [true, true, false, false]]

Any function that takes a vector argument can be lifted to an operation on block vectors.

q = lift_to_block(length)
#-> lift_to_block(length)

q(@VectorTree [String] [["GARRY M", "ANTHONY R", "DANA A"], ["JOSE S", "CHARLES S"]])
#-> [3, 2]

Some vector functions may expect a non-empty vector as an argument. In this case, we should provide the value to replace empty blocks.

q = lift_to_block(maximum, missing)
#-> lift_to_block(maximum, missing)

q(@VectorTree [Int] [[260004, 185364, 170112], [], [202728, 197736]])
#-> Union{Missing, Int}[260004, missing, 202728]

Decoding vectors

Any vector of tuples can be converted to a tuple vector.

q = decode_tuple()
#-> decode_tuple()

q([("GARRY M", 260004), ("ANTHONY R", 185364), ("DANA A", 170112)]) |> display
#=>
TupleVector of 3 × (String, Int):
 ("GARRY M", 260004)
 ("ANTHONY R", 185364)
 ("DANA A", 170112)
=#

Vectors of named tuples are also supported.

q([(name="GARRY M", salary=260004), (name="ANTHONY R", salary=185364), (name="DANA A", salary=170112)]) |> display
#=>
TupleVector of 3 × (name = String, salary = Int):
 (name = "GARRY M", salary = 260004)
 (name = "ANTHONY R", salary = 185364)
 (name = "DANA A", salary = 170112)
=#

A vector of vector objects can be converted to a block vector.

q = decode_vector()
#-> decode_vector()

q([[260004, 185364, 170112], Int[], [202728, 197736]])
#-> @VectorTree [Int] [[260004, 185364, 170112], [], [202728, 197736]]

Similarly, a vector containing missing values can be converted to a block vector with zero- and one-element blocks.

q = decode_missing()
#-> decode_missing()

q([260004, 185364, 170112, missing, 202728, 197736])
#-> @VectorTree [Int, OPT] [260004, 185364, 170112, missing, 202728, 197736]

Tuple vectors

To create a tuple vector, we use the combinator tuple_of(). Its arguments are the functions that generate the columns of the tuple.

q = tuple_of(:title => lift(titlecase), :last => lift(last))
#-> tuple_of([:title, :last], [lift(titlecase), lift(last)])

q(["GARRY M", "ANTHONY R", "DANA A"]) |> display
#=>
TupleVector of 3 × (title = String, last = Char):
 (title = "Garry M", last = 'M')
 (title = "Anthony R", last = 'R')
 (title = "Dana A", last = 'A')
=#

To extract a column of a tuple vector, we use the primitive column(). It accepts either the column position or the column name.

q = column(1)
#-> column(1)

q(@VectorTree (name = String, salary = Int) ["GARRY M" 260004; "ANTHONY R" 185364; "DANA A" 170112])
#-> ["GARRY M", "ANTHONY R", "DANA A"]

q = column(:salary)
#-> column(:salary)

q(@VectorTree (name = String, salary = Int) ["GARRY M" 260004; "ANTHONY R" 185364; "DANA A" 170112])
#-> [260004, 185364, 170112]

Finally, we can apply an arbitrary transformation to a selected column of a tuple vector.

q = in_tuple(:name, lift(titlecase))
#-> in_tuple(:name, lift(titlecase))

q(@VectorTree (name = String, salary = Int) ["GARRY M" 260004; "ANTHONY R" 185364; "DANA A" 170112]) |> display
#=>
TupleVector of 3 × (name = String, salary = Int):
 (name = "Garry M", salary = 260004)
 (name = "Anthony R", salary = 185364)
 (name = "Dana A", salary = 170112)
=#

Block vectors

Primitive as_block() wraps the elements of the input vector to one-element blocks.

q = as_block()
#-> as_block()

q(["GARRY M", "ANTHONY R", "DANA A"])
#-> @VectorTree [String, REG] ["GARRY M", "ANTHONY R", "DANA A"]

In the opposite direction, primitive flat_block() flattens a block vector with block elements.

q = flat_block()
#-> flat_block()

q(@VectorTree [[String]] [[["GARRY M"], ["ANTHONY R", "DANA A"]], [missing, ["JOSE S"], ["CHARLES S"]]])
#-> @VectorTree [String] [["GARRY M", "ANTHONY R", "DANA A"], ["JOSE S", "CHARLES S"]]

Finally, we can apply an arbitrary transformation to every element of a block vector.

q = in_block(lift(titlecase))
#-> in_block(lift(titlecase))

q(@VectorTree [String] [["GARRY M", "ANTHONY R", "DANA A"], ["JOSE S", "CHARLES S"]])
#-> @VectorTree [String] [["Garry M", "Anthony R", "Dana A"], ["Jose S", "Charles S"]]

The pull_block() primitive converts a tuple vector with a block column to a block vector of tuples.

q = pull_block(1)
#-> pull_block(1)

q(@VectorTree ([Int], [Int]) [
    [260004, 185364, 170112]    200000
    missing                     200000
    [202728, 197736]            [200000, 200000]]
) |> display
#=>
BlockVector of 3 × [(Int, [Int])]:
 [(260004, [200000]), (185364, [200000]), (170112, [200000])]
 []
 [(202728, [200000, 200000]), (197736, [200000, 200000])]
=#

It is also possible to pull all block columns from a tuple vector.

q = pull_every_block()
#-> pull_every_block()

q(@VectorTree ([Int], [Int]) [
    [260004, 185364, 170112]    200000
    missing                     200000
    [202728, 197736]            [200000, 200000]]
) |> display
#=>
BlockVector of 3 × [(Int, Int)]:
 [(260004, 200000), (185364, 200000), (170112, 200000)]
 []
 [(202728, 200000), (202728, 200000), (197736, 200000), (197736, 200000)]
=#

Composition

We can compose a sequence of transformations using the chain_of() combinator.

q = chain_of(
        column(:employee),
        in_block(lift(titlecase)))
#-> chain_of(column(:employee), in_block(lift(titlecase)))

q(@VectorTree (department = String, employee = [String]) [
    "POLICE"    ["GARRY M", "ANTHONY R", "DANA A"]
    "FIRE"      ["JOSE S", "CHARLES S"]])
#-> @VectorTree [String] [["Garry M", "Anthony R", "Dana A"], ["Jose S", "Charles S"]]

The empty chain chain_of() has an alias pass().

q = pass()
#-> pass()

q(["GARRY M", "ANTHONY R", "DANA A"])
#-> ["GARRY M", "ANTHONY R", "DANA A"]