Reference

Reference

DataKnots.It — Constant.

It :: AbstractQuery

In a query expression, use It to refer to the query's input.

julia> DataKnot(3)[It .+ 1]
│ It │
┼────┼
│  4 │

It is the identity with respect to query composition.

julia> DataKnot()[Lift('a':'c') >> It]
  │ It │
──┼────┼
1 │ a  │
2 │ b  │
3 │ c  │

It provides a shorthand notation for data navigation using Get, so that It.a.x is equivalent to Get(:a) >> Get(:x).

julia> DataKnot((a=(x=1,y=2),))[It.a]
│ a    │
│ x  y │
┼──────┼
│ 1  2 │

julia> DataKnot((a=(x=1,y=2),))[It.a.x]
│ x │
┼───┼
│ 1 │

source

DataKnots.DataKnot — Type.

DataKnot(::Any=nothing)
DataKnot(::AbstactVector, card=:x0toN)

The DataKnot constructor wraps a given value so that it could be used to start a query. When the argument is an vector, we can clarify its cardinality (:x0to1, :x1to1, :x0toN, :x1toN).

The simplest knot holds the value nothing.

julia> DataKnot()
│ It │
┼────┼
│    │

An empty knot can be constructed with missing.

julia> DataKnot(missing)
│ It │
┼────┼

It's often useful to wrap a dataset in a one-field tuple.

julia> DataKnot((dataset='a':'c',))
│ dataset │
┼─────────┼
│ a; b; c │

By default, vectors have no cardinality constraints. To indicate that vector will always has at least one element, use :x1toN.

julia> DataKnot('a':'c', :x1toN)
  │ It │
──┼────┼
1 │ a  │
2 │ b  │
3 │ c  │

get(::DataKnot)

Use get to extract the underlying value held by a knot.

julia> get(DataKnot("Hello World"))
"Hello World"

getindex(::DataKnot, X; kwargs...)

We can query a knot using array indexing notation.

julia> DataKnot((dataset='a':'c',))[Count(It.dataset)]
│ It │
┼────┼
│  3 │

Query parameters are provided as keyword arguments.

julia> DataKnot(1:3)[PWR=2, It .^ It.PWR]
  │ It │
──┼────┼
1 │  1 │
2 │  4 │
3 │  9 │

source

DataKnots.Count — Method.

Count(X) :: Query

In the combinator form, Count(X) emits the number of elements produced by X.

julia> X = Lift('a':'c');

julia> DataKnot()[Count(X)]
│ It │
┼────┼
│  3 │

Each(X >> Count) :: Query

In the query form, Count emits the number of elements in its input.

julia> X = Lift('a':'c');

julia> DataKnot()[X >> Count]
│ It │
┼────┼
│  3 │

To limit the scope of aggregation, use Each.

julia> X = Lift('a':'c');

julia> DataKnot()[Lift(1:3) >> Each(X >> Count)]
  │ It │
──┼────┼
1 │  3 │
2 │  3 │
3 │  3 │

source

DataKnots.Drop — Method.

Drop(N) :: Query

This query drops the first N elements of its input, preserving the rest.

julia> DataKnot()[Lift('a':'c') >> Drop(2)]
  │ It │
──┼────┼
1 │ c  │

Drop(-N) takes the last N elements.

julia> DataKnot()[Lift('a':'c') >> Drop(-2)]
  │ It │
──┼────┼
1 │ b  │
2 │ c  │

source

DataKnots.Each — Method.

Each(X) :: Query

This evaluates X elementwise.

julia> X = Lift('a':'c') >> Count;

julia> DataKnot()[Lift(1:3) >> Each(X)]
  │ It │
──┼────┼
1 │  3 │
2 │  3 │
3 │  3 │

Compare this with the query without Each.

julia> X = Lift('a':'c') >> Count;

julia> DataKnot()[Lift(1:3) >> X]
│ It │
┼────┼
│  9 │

source

DataKnots.Filter — Method.

Filter(X) :: Query

This query emits the elements from its input that satisfy a given condition.

julia> DataKnot(1:5)[Filter(isodd.(It))]
  │ It │
──┼────┼
1 │  1 │
2 │  3 │
3 │  5 │

When the predicate query produces an empty output, the condition is presumed to have failed.

julia> DataKnot('a':'c')[Filter(missing)]
│ It │
┼────┼

When the predicate produces plural output, the condition succeeds if at least one output value is true.

julia> DataKnot('a':'c')[Filter([true,false])]
  │ It │
──┼────┼
1 │ a  │
2 │ b  │
3 │ c  │

source

DataKnots.Get — Method.

Get(lbl::Symbol) :: Query

This query extracts a field value by its label.

julia> DataKnot((x=1, y=2))[Get(:x)]
│ x │
┼───┼
│ 1 │

This has a shorthand form using It.

julia> DataKnot((x=1, y=2))[It.x]
│ x │
┼───┼
│ 1 │

With unlabeled fields, ordinal labels (A, B, ...) can be used.

julia> DataKnot((1,2))[It.B]
│ It │
┼────┼
│  2 │

source

DataKnots.Given — Method.

Given(X₁, X₂ … Xₙ, Q) :: Query

This evaluates Q in a context augmented with named parameters added by a set of queries.

julia> DataKnot()[Given(:x => 2, It.x .+ 1)]
│ It │
┼────┼
│  3 │

source

DataKnots.Keep — Method.

Keep(X₁, X₂ … Xₙ) :: Query

Keep evaluates named queries, making their results available for subsequent computation.

julia> DataKnot()[Keep(:x => 2) >> It.x]
│ x │
┼───┼
│ 2 │

Keep does not otherwise change its input.

julia> DataKnot(1)[Keep(:x => 2) >> (It .+ It.x)]
│ It │
┼────┼
│  3 │

source

DataKnots.Label — Method.

Label(lbl::Symbol) :: Query

This assigns a label to the output.

julia> DataKnot()[Lift("Hello World") >> Label(:greeting)]
│ greeting    │
┼─────────────┼
│ Hello World │

A label could also be assigned using the => operator.

julia> DataKnot()[:greeting => Lift("Hello World")]
│ greeting    │
┼─────────────┼
│ Hello World │

source

DataKnots.Lift — Method.

Lift(f, (X₁, X₂ … Xₙ)) :: Query

Lift lets you use a function as a query combinator.

julia> DataKnot((x=1, y=2))[Lift(+, (It.x, It.y))]
│ It │
┼────┼
│  3 │

Lift is implicitly used when a function is broadcast over queries.

julia> DataKnot((x=1, y=2))[It.x .+ It.y]
│ It │
┼────┼
│  3 │

Functions accepting a AbstractVector can be used with plural queries.

julia> DataKnot()[sum.(Lift(1:3))]
│ It │
┼────┼
│  6 │

Functions returning AbstractVector become plural queries.

julia> DataKnot((x='a', y='c'))[Lift(:, (It.x, It.y))]
  │ It │
──┼────┼
1 │ a  │
2 │ b  │
3 │ c  │

source

DataKnots.Lift — Method.

Lift(val) :: Query

This converts any value to a constant query.

julia> DataKnot()[Lift("Hello")]
│ It    │
┼───────┼
│ Hello │

AbstractVector objects become plural queries.

julia> DataKnot()[Lift('a':'c')]
  │ It │
──┼────┼
1 │ a  │
2 │ b  │
3 │ c  │

The missing value makes an query with no output.

julia> DataKnot()[Lift(missing)]
│ It │
┼────┼

source

DataKnots.Max — Method.

 Max(X) :: Query

In the combinator form, Max(X) finds the maximum among the elements produced by X.

julia> X = Lift(1:3);

julia> DataKnot()[Max(X)]
│ It │
┼────┼
│  3 │

The Max of an empty input is empty.

julia> DataKnot()[Max(Int[])]
│ It │
┼────┼

Each(X >> Max) :: Query

In the query form, Max finds the maximum of its input elements.

julia> X = Lift(1:3);

julia> DataKnot()[X >> Max]
│ It │
┼────┼
│  3 │

source

DataKnots.Min — Method.

 Min(X) :: Query

In the combinator form, Min(X) finds the minimum among the elements produced by X.

julia> X = Lift(1:3);

julia> DataKnot()[Min(X)]
│ It │
┼────┼
│  1 │

The Min of an empty input is empty.

julia> DataKnot()[Min(Int[])]
│ It │
┼────┼

Each(X >> Min) :: Query

In the query form, Min finds the minimum of its input elements.

julia> X = Lift(1:3);

julia> DataKnot()[X >> Min]
│ It │
┼────┼
│  1 │

source

DataKnots.Record — Method.

Record(X₁, X₂ … Xₙ) :: Query

Record(X₁, X₂ … Xₙ) emits records whose fields are generated by X₁, X₂ … Xₙ.

julia> DataKnot()[Lift(1:3) >> Record(It, It .* It)]
  │ #A  #B │
──┼────────┼
1 │  1   1 │
2 │  2   4 │
3 │  3   9 │

Field labels are inherited from queries.

julia> DataKnot()[Lift(1:3) >> Record(:x => It,
                                      :x² => It .* It)]
  │ x  x² │
──┼───────┼
1 │ 1   1 │
2 │ 2   4 │
3 │ 3   9 │

source

DataKnots.Sum — Method.

Sum(X) :: Query

In the combinator form, Sum(X) emits the sum of elements produced by X.

julia> X = Lift(1:3);

julia> DataKnot()[Sum(X)]
│ It │
┼────┼
│  6 │

The Sum of an empty input is 0.

julia> DataKnot()[Sum(Int[])]
│ It │
┼────┼
│  0 │

Each(X >> Sum) :: Query

In the query form, Sum emits the sum of input elements.

julia> X = Lift(1:3);

julia> DataKnot()[X >> Sum]
│ It │
┼────┼
│  6 │

source

DataKnots.Tag — Method.

Tag(name::Symbol, F) :: Query

This provides a substitute name for a query.

julia> IncIt = It .+ 1
It .+ 1

julia> IncIt = Tag(:IncIt, It .+ 1)
IncIt

Tag(name::Symbol, (X₁, X₂ … Xₙ), F) :: Query

This provides a substitute name for a query combinator.

julia> Inc(X) = Lift(+, (X, 1));

julia> Inc(It)
Lift(+, (It, 1))

julia> Inc(X) = Tag(:Inc, (X,), Lift(+, (X, 1)));

julia> Inc(It)
Inc(It)

source

DataKnots.Take — Method.

Take(N) :: Query

This query preserves the first N elements of its input, dropping the rest.

julia> DataKnot()[Lift('a':'c') >> Take(2)]
  │ It │
──┼────┼
1 │ a  │
2 │ b  │

Take(-N) drops the last N elements.

julia> DataKnot()[Lift('a':'c') >> Take(-2)]
  │ It │
──┼────┼
1 │ a  │

source